Keep Hermes hook installer split out
